requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
pytesseract>=0.3.10
pyserial>=3.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0 